move fetch state to child component
